@@ -1,6 +1,6 @@
 {
   "name": "tcomb-form-native",
-  "version": "0.5.1",
+  "version": "0.5.2",
   "description": "react-native powered UI library for developing forms writing less code",
   "main": "index.js",
   "scripts": {
@@ -25,8 +25,7 @@
     "babel-core": "5.8.34",
     "eslint": "0.23.0",
     "eslint-plugin-react": "2.5.2",
-    "mockery": "1.4.0",
-    "react": "0.14.7",
+    "react": "^15.0.0",
     "tape": "3.5.0"
   },
   "tags": [
