@@ -1,9 +1,7 @@
-'use strict';
+import t, { mixin } from 'tcomb-validation';
 
-var t = require('tcomb-validation');
-
-function getOptionsOfEnum(type) {
-  var enums = type.meta.map;
+export function getOptionsOfEnum(type) {
+  const enums = type.meta.map;
   return Object.keys(enums).map(value => {
     return {
       value,
@@ -12,13 +10,12 @@ function getOptionsOfEnum(type) {
   });
 }
 
-function getTypeInfo(type) {
-
-  var innerType = type;
-  var isMaybe = false;
-  var isSubtype = false;
-  var kind;
-  var innerGetValidationErrorMessage;
+export function getTypeInfo(type) {
+  let innerType = type;
+  let isMaybe = false;
+  let isSubtype = false;
+  let kind;
+  let innerGetValidationErrorMessage;
 
   while (innerType) {
     kind = innerType.meta.kind;
@@ -38,10 +35,10 @@ function getTypeInfo(type) {
     break;
   }
 
-  var getValidationErrorMessage = innerGetValidationErrorMessage ? function (value, path, context) {
-    var result = t.validate(value, type, {path, context});
+  const getValidationErrorMessage = innerGetValidationErrorMessage ? (value, path, context) => {
+    const result = t.validate(value, type, {path, context});
     if (!result.isValid()) {
-      for (var i = 0, len = result.errors.length; i < len; i++ ) {
+      for (let i = 0, len = result.errors.length; i < len; i++ ) {
         if (t.Function.is(result.errors[i].expected.getValidationErrorMessage)) {
           return result.errors[i].message;
         }
@@ -61,26 +58,106 @@ function getTypeInfo(type) {
 
 // thanks to https://github.com/epeli/underscore.string
 
-function underscored(s){
+function underscored(s) {
   return s.trim().replace(/([a-z\d])([A-Z]+)/g, '$1_$2').replace(/[-\s]+/g, '_').toLowerCase();
 }
 
-function capitalize(s){
+function capitalize(s) {
   return s.charAt(0).toUpperCase() + s.slice(1);
 }
 
-function humanize(s){
+export function humanize(s) {
   return capitalize(underscored(s).replace(/_id$/, '').replace(/_/g, ' '));
 }
 
-function merge(a, b) {
-  return t.mixin(t.mixin({}, a), b, true);
+export function merge(a, b) {
+  return mixin(mixin({}, a), b, true);
 }
 
-module.exports = {
-  getOptionsOfEnum,
-  getTypeInfo,
-  humanize,
-  merge
-};
+export function move(arr, fromIndex, toIndex) {
+  const element = arr.splice(fromIndex, 1)[0];
+  arr.splice(toIndex, 0, element);
+  return arr;
+}
+
+export class UIDGenerator {
+
+  constructor(seed) {
+    this.seed = 'tfid-' + seed + '-';
+    this.counter = 0;
+  }
+
+  next() {
+    return this.seed + (this.counter++); // eslint-disable-line space-unary-ops
+  }
+
+}
+
+function containsUnion(type) {
+  switch (type.meta.kind) {
+  case 'union' :
+    return true;
+  case 'maybe' :
+  case 'subtype' :
+    return containsUnion(type.meta.type);
+  default :
+    return false;
+  }
+}
+
+function getUnionConcreteType(type, value) {
+  const kind = type.meta.kind;
+  if (kind === 'union') {
+    const concreteType = type.dispatch(value);
+    if (process.env.NODE_ENV !== 'production') {
+      t.assert(t.isType(concreteType), () => 'Invalid value ' + t.assert.stringify(value) + ' supplied to ' + t.getTypeName(type) + ' (no constructor returned by dispatch)' );
+    }
+    return concreteType;
+  } else if (kind === 'maybe') {
+    return t.maybe(getUnionConcreteType(type.meta.type, value), type.meta.name);
+  } else if (kind === 'subtype') {
+    return t.subtype(getUnionConcreteType(type.meta.type, value), type.meta.predicate, type.meta.name);
+  }
+}
+
+export function getTypeFromUnion(type, value) {
+  if (containsUnion(type)) {
+    return getUnionConcreteType(type, value);
+  }
+  return type;
+}
+
+function getUnion(type) {
+  if (type.meta.kind === 'union') {
+    return type;
+  }
+  return getUnion(type.meta.type);
+}
+
+function findIndex(arr, element) {
+  for (let i = 0, len = arr.length; i < len; i++ ) {
+    if (arr[i] === element) {
+      return i;
+    }
+  }
+  return -1;
+}
+
+export function getComponentOptions(options, defaultOptions, value, type) {
+  if (t.Nil.is(options)) {
+    return defaultOptions;
+  }
+  if (t.Function.is(options)) {
+    return options(value);
+  }
+  if (t.Array.is(options) && containsUnion(type)) {
+    const union = getUnion(type);
+    const concreteType = union.dispatch(value);
+    const index = findIndex(union.meta.types, concreteType);
+    // recurse
+    return getComponentOptions(options[index], defaultOptions, value, concreteType);
+  }
+  return options;
+}
+
 
