@@ -1,19 +1,31 @@
-'use strict';
+import React from 'react';
+import t from 'tcomb-validation';
+import {
+  humanize,
+  merge,
+  getTypeInfo,
+  getOptionsOfEnum,
+  move,
+  UIDGenerator,
+  getTypeFromUnion,
+  getComponentOptions
+} from './util';
 
-var React = require('react');
-var t = require('tcomb-validation');
-var { humanize, merge, getTypeInfo, getOptionsOfEnum } = require('./util');
+const SOURCE = 'tcomb-form-native';
+const nooptions = Object.freeze({});
+const noop = function () {};
+const noobj = Object.freeze({});
+const noarr = Object.freeze([]);
+const Nil = t.Nil;
 
-var SOURCE = 'tcomb-form-native';
-var nooptions = Object.freeze({});
-var noop = function () {};
-var Nil = t.Nil;
-
-function getComponent(type, options) {
+function getFormComponent(type, options) {
   if (options.factory) {
     return options.factory;
   }
-  var name = t.getTypeName(type);
+  if (type.getTcombFormFactory) {
+    return type.getTcombFormFactory(options);
+  }
+  const name = t.getTypeName(type);
   switch (type.meta.kind) {
     case 'irreducible' :
       return (
@@ -23,11 +35,13 @@ function getComponent(type, options) {
       );
     case 'struct' :
       return Struct;
+    case 'list' :
+      return List;
     case 'enums' :
       return Select;
     case 'maybe' :
     case 'subtype' :
-      return getComponent(type.meta.type, options);
+      return getFormComponent(type.meta.type, options);
     default :
       t.fail(`[${SOURCE}] unsupported type ${name}`);
   }
@@ -60,7 +74,7 @@ class Component extends React.Component {
   }
 
   shouldComponentUpdate(nextProps, nextState) {
-    var should = (
+    const should = (
       nextState.value !== this.state.value ||
       nextState.hasError !== this.state.hasError ||
       nextProps.options !== this.props.options ||
@@ -100,7 +114,7 @@ class Component extends React.Component {
   }
 
   validate() {
-    var result = t.validate(this.getValue(), this.props.type, this.getValidationOptions());
+    const result = t.validate(this.getValue(), this.props.type, this.getValidationOptions());
     this.setState({hasError: !result.isValid()});
     return result;
   }
@@ -114,17 +128,14 @@ class Component extends React.Component {
   }
 
   getDefaultLabel() {
-    var ctx = this.props.ctx;
+    const ctx = this.props.ctx;
     if (ctx.label) {
       return ctx.label + (this.typeInfo.isMaybe ? this.getI18n().optional : this.getI18n().required);
     }
   }
 
   getLabel() {
-    var ctx = this.props.ctx;
-    var legend = this.props.options.legend;
-    var label = this.props.options.label;
-    label = label || legend;
+    let label = this.props.options.label || this.props.options.legend;
     if (Nil.is(label) && this.getAuto() === 'labels') {
       label = this.getDefaultLabel();
     }
@@ -133,9 +144,9 @@ class Component extends React.Component {
 
   getError() {
     if (this.hasError()) {
-      var error = this.props.options.error || this.typeInfo.getValidationErrorMessage;
+      const error = this.props.options.error || this.typeInfo.getValidationErrorMessage;
       if (t.Function.is(error)) {
-        var validationOptions = this.getValidationOptions();
+        const validationOptions = this.getValidationOptions();
         return error(this.getValue(), validationOptions.path, validationOptions.context);
       }
       return error;
@@ -169,10 +180,10 @@ class Component extends React.Component {
   }
 
   render() {
-    var locals = this.getLocals();
+    const locals = this.getLocals();
     // getTemplate is the only required implementation when extending Component
     t.assert(t.Function.is(this.getTemplate), `[${SOURCE}] missing getTemplate method of component ${this.constructor.name}`);
-    var template = this.getTemplate();
+    const template = this.getTemplate();
     return template(locals);
   }
 
@@ -188,15 +199,15 @@ function toNull(value) {
 }
 
 function parseNumber(value) {
-  var n = parseFloat(value);
-  var isNumeric = (value - n + 1) >= 0;
+  const n = parseFloat(value);
+  const isNumeric = (value - n + 1) >= 0;
   return isNumeric ? n : toNull(value);
 }
 
 class Textbox extends Component {
 
   getTransformer() {
-    var options = this.props.options;
+    const options = this.props.options;
     return options.transformer ? options.transformer :
       this.typeInfo.innerType === t.Number ? Textbox.numberTransformer :
       Textbox.transformer;
@@ -207,7 +218,7 @@ class Textbox extends Component {
   }
 
   getPlaceholder() {
-    var placeholder = this.props.options.placeholder;
+    let placeholder = this.props.options.placeholder;
     if (Nil.is(placeholder) && this.getAuto() === 'placeholders') {
       placeholder = this.getDefaultLabel();
     }
@@ -215,7 +226,7 @@ class Textbox extends Component {
   }
 
   getLocals() {
-    var locals = super.getLocals();
+    const locals = super.getLocals();
     locals.placeholder = this.getPlaceholder();
     locals.onChangeNative = this.props.options.onChange;
 
@@ -272,7 +283,7 @@ class Checkbox extends Component {
   }
 
   getLocals() {
-    var locals = super.getLocals();
+    const locals = super.getLocals();
     // checkboxes must always have a label
     locals.label = locals.label || this.getDefaultLabel();
 
@@ -297,7 +308,7 @@ Checkbox.transformer = {
 class Select extends Component {
 
   getTransformer() {
-    var options = this.props.options;
+    const options = this.props.options;
     if (options.transformer) {
       return options.transformer;
     }
@@ -317,12 +328,12 @@ class Select extends Component {
   }
 
   getOptions() {
-    var options = this.props.options;
-    var items = options.options ? options.options.slice() : getOptionsOfEnum(this.getEnum());
+    const options = this.props.options;
+    const items = options.options ? options.options.slice() : getOptionsOfEnum(this.getEnum());
     if (options.order) {
       items.sort(getComparator(options.order));
     }
-    var nullOption = this.getNullOption();
+    const nullOption = this.getNullOption();
     if (options.nullOption !== false) {
       items.unshift(nullOption);
     }
@@ -330,7 +341,7 @@ class Select extends Component {
   }
 
   getLocals() {
-    var locals = super.getLocals();
+    const locals = super.getLocals();
     locals.options = this.getOptions();
 
     [
@@ -360,7 +371,7 @@ class DatePicker extends Component {
   }
 
   getLocals() {
-    var locals = super.getLocals();
+    const locals = super.getLocals();
 
     [
       'help',
@@ -393,26 +404,25 @@ class Struct extends Component {
   }
 
   getValue() {
-    var value = {};
-    for (var ref in this.refs) {
+    const value = {};
+    for (const ref in this.refs) {
       value[ref] = this.refs[ref].getValue();
     }
     return this.getTransformer().parse(value);
   }
 
   validate() {
-
-    var value = {};
-    var errors = [];
-    var hasError = false;
-    var result;
+    let value = {};
+    let errors = [];
+    let hasError = false;
+    let result;
 
     if (this.typeInfo.isMaybe && this.isValueNully()) {
       this.removeErrors();
       return new t.ValidationResult({errors: [], value: null});
     }
 
-    for (var ref in this.refs) {
+    for (const ref in this.refs) {
       if (this.refs.hasOwnProperty(ref)) {
         result = this.refs[ref].validate();
         errors = errors.concat(result.errors);
@@ -421,7 +431,7 @@ class Struct extends Component {
     }
 
     if (errors.length === 0) {
-      var InnerType = this.typeInfo.innerType;
+      const InnerType = this.typeInfo.innerType;
       value = new InnerType(value);
       if (this.typeInfo.isSubtype && errors.length === 0) {
         result = t.validate(value, this.props.type, this.getValidationOptions());
@@ -435,7 +445,7 @@ class Struct extends Component {
   }
 
   onChange(fieldName, fieldValue, path) {
-    var value = t.mixin({}, this.state.value);
+    const value = t.mixin({}, this.state.value);
     value[fieldName] = fieldValue;
     this.setState({value}, () => {
       this.props.onChange(value, path);
@@ -450,10 +460,6 @@ class Struct extends Component {
     return this.props.options.template || this.getTemplates().struct;
   }
 
-  getStylesheet() {
-    return this.props.options.stylesheet || this.props.ctx.stylesheet;
-  }
-
   getTypeProps() {
     return this.typeInfo.innerType.meta.props;
   }
@@ -463,20 +469,23 @@ class Struct extends Component {
   }
 
   getInputs() {
-    var { ctx, options } = this.props;
-    var props = this.getTypeProps();
-    var auto = this.getAuto();
-    var i18n =  this.getI18n();
-    var config = this.getConfig();
-    var value = this.state.value || {};
-    var templates = this.getTemplates();
-    var stylesheet = this.getStylesheet();
-    var inputs = {};
-    for (var prop in props) {
+    const { ctx, options } = this.props;
+    const props = this.getTypeProps();
+    const auto = this.getAuto();
+    const i18n =  this.getI18n();
+    const config = this.getConfig();
+    const value = this.state.value || {};
+    const templates = this.getTemplates();
+    const stylesheet = this.getStylesheet();
+    const inputs = {};
+    for (const prop in props) {
       if (props.hasOwnProperty(prop)) {
-        var propType = props[prop];
-        var propOptions = options.fields && options.fields[prop] ? options.fields[prop] : nooptions;
-        inputs[prop] = React.createElement(getComponent(propType, propOptions), {
+        const type = props[prop];
+        const propValue = value[prop];
+        const propType = getTypeFromUnion(type, propValue);
+        const fieldsOptions = options.fields || noobj;
+        const propOptions = getComponentOptions(fieldsOptions[prop], noobj, propValue, type);
+        inputs[prop] = React.createElement(getFormComponent(propType, propOptions), {
           key: prop,
           ref: prop,
           type: propType,
@@ -485,6 +494,7 @@ class Struct extends Component {
           onChange: this.onChange.bind(this, prop),
           ctx: {
             context: ctx.context,
+            uidGenerator: ctx.uidGenerator,
             auto,
             config,
             label: humanize(prop),
@@ -500,8 +510,8 @@ class Struct extends Component {
   }
 
   getLocals() {
-    var templates = this.getTemplates();
-    var locals = super.getLocals();
+    const templates = this.getTemplates();
+    const locals = super.getLocals();
     locals.order = this.getOrder();
     locals.inputs = this.getInputs();
     locals.template = templates.struct;
@@ -510,6 +520,218 @@ class Struct extends Component {
 
 }
 
+function toSameLength(value, keys, uidGenerator) {
+  if (value.length === keys.length) {
+    return keys;
+  }
+  const ret = [];
+  for (let i = 0, len = value.length; i < len; i++ ) {
+    ret[i] = keys[i] || uidGenerator.next();
+  }
+  return ret;
+}
+
+export class List extends Component {
+
+  constructor(props) {
+    super(props);
+    this.state.keys = this.state.value.map(() => props.ctx.uidGenerator.next());
+  }
+
+  componentWillReceiveProps(props) {
+    if (props.type !== this.props.type) {
+      this.typeInfo = getTypeInfo(props.type);
+    }
+    const value = this.getTransformer().format(props.value);
+    this.setState({
+      value,
+      keys: toSameLength(value, this.state.keys, props.ctx.uidGenerator)
+    });
+  }
+
+  isValueNully() {
+    return this.state.value.length === 0;
+  }
+
+  removeErrors() {
+    this.setState({ hasError: false });
+    Object.keys(this.refs).forEach((ref) => this.refs[ref].removeErrors());
+  }
+
+  getValue() {
+    const value = [];
+    for (let i = 0, len = this.state.value.length; i < len; i++ ) {
+      if (this.refs.hasOwnProperty(i)) {
+        value.push(this.refs[i].getValue());
+      }
+    }
+    return this.getTransformer().parse(value);
+  }
+
+  validate() {
+    const value = [];
+    let errors = [];
+    let hasError = false;
+    let result;
+
+    if (this.typeInfo.isMaybe && this.isValueNully()) {
+      this.removeErrors();
+      return new t.ValidationResult({errors: [], value: null});
+    }
+
+    for (let i = 0, len = this.state.value.length; i < len; i++ ) {
+      result = this.refs[i].validate();
+      errors = errors.concat(result.errors);
+      value.push(result.value);
+    }
+
+    // handle subtype
+    if (this.typeInfo.isSubtype && errors.length === 0) {
+      result = t.validate(value, this.props.type, this.getValidationOptions());
+      hasError = !result.isValid();
+      errors = errors.concat(result.errors);
+    }
+
+    this.setState({hasError: hasError});
+    return new t.ValidationResult({errors: errors, value: value});
+  }
+
+  onChange(value, keys, path, kind) {
+    const allkeys = toSameLength(value, keys, this.props.ctx.uidGenerator);
+    this.setState({ value, keys: allkeys, isPristine: false }, () => {
+      this.props.onChange(value, path, kind);
+    });
+  }
+
+  addItem() {
+    const value = this.state.value.concat(undefined);
+    const keys = this.state.keys.concat(this.props.ctx.uidGenerator.next());
+    this.onChange(value, keys, this.props.ctx.path.concat(value.length - 1), 'add');
+  }
+
+  onItemChange(itemIndex, itemValue, path, kind) {
+    const value = this.state.value.slice();
+    value[itemIndex] = itemValue;
+    this.onChange(value, this.state.keys, path, kind);
+  }
+
+  removeItem(i) {
+    const value = this.state.value.slice();
+    value.splice(i, 1);
+    const keys = this.state.keys.slice();
+    keys.splice(i, 1);
+    this.onChange(value, keys, this.props.ctx.path.concat(i), 'remove');
+  }
+
+  moveUpItem(i) {
+    if (i > 0) {
+      this.onChange(
+        move(this.state.value.slice(), i, i - 1),
+        move(this.state.keys.slice(), i, i - 1),
+        this.props.ctx.path.concat(i),
+        'moveUp'
+      );
+    }
+  }
+
+  moveDownItem(i) {
+    if (i < this.state.value.length - 1) {
+      this.onChange(
+        move(this.state.value.slice(), i, i + 1),
+        move(this.state.keys.slice(), i, i + 1),
+        this.props.ctx.path.concat(i),
+        'moveDown'
+      );
+    }
+  }
+
+  getTemplates() {
+    return merge(this.props.ctx.templates, this.props.options.templates);
+  }
+
+  getTemplate() {
+    return this.props.options.template || this.getTemplates().list;
+  }
+
+  getItems() {
+    const { options, ctx } = this.props;
+    const auto = this.getAuto();
+    const i18n = this.getI18n();
+    const config = this.getConfig();
+    const stylesheet = this.getStylesheet();
+    const templates = this.getTemplates();
+    const value = this.state.value;
+    return value.map((itemValue, i) => {
+      const type = this.typeInfo.innerType.meta.type;
+      const itemType = getTypeFromUnion(type, itemValue);
+      const itemOptions = getComponentOptions(options.item, noobj, itemValue, type);
+      const ItemComponent = getFormComponent(itemType, itemOptions);
+      const buttons = [];
+      if (!options.disableRemove) {
+        buttons.push({
+          type: 'remove',
+          label: i18n.remove,
+          click: this.removeItem.bind(this, i)
+        });
+      }
+      if (!options.disableOrder) {
+        buttons.push({
+          type: 'move-up',
+          label: i18n.up,
+          click: this.moveUpItem.bind(this, i)
+        },
+        {
+          type: 'move-down',
+          label: i18n.down,
+          click: this.moveDownItem.bind(this, i)
+        });
+      }
+      return {
+        input: React.createElement(ItemComponent, {
+          ref: i,
+          type: itemType,
+          options: itemOptions,
+          value: itemValue,
+          onChange: this.onItemChange.bind(this, i),
+          ctx: {
+            context: ctx.context,
+            uidGenerator: ctx.uidGenerator,
+            auto,
+            config,
+            label: ctx.label ? `${ctx.label}[${i + 1}]` : String(i + 1),
+            i18n,
+            stylesheet,
+            templates,
+            path: ctx.path.concat(i)
+          }
+        }),
+        key: this.state.keys[i],
+        buttons: buttons
+      };
+    });
+  }
+
+  getLocals() {
+    const options = this.props.options;
+    const i18n = this.getI18n();
+    const locals = super.getLocals();
+    locals.add = options.disableAdd ? null : {
+      type: 'add',
+      label: i18n.add,
+      click: this.addItem.bind(this)
+    };
+    locals.items = this.getItems();
+    locals.className = options.className;
+    return locals;
+  }
+
+}
+
+List.transformer = {
+  format: value => Nil.is(value) ? noarr : value,
+  parse: value => value
+};
+
 class Form extends React.Component {
 
   validate() {
@@ -517,7 +739,7 @@ class Form extends React.Component {
   }
 
   getValue() {
-    var result = this.validate();
+    const result = this.validate();
     return result.isValid() ? result.value : null;
   }
 
@@ -526,23 +748,37 @@ class Form extends React.Component {
     return path.reduce((input, name) => input.refs[name], this.refs.input);
   }
 
+  getUIDGenerator() {
+    const seed = this._reactInternalInstance && this._reactInternalInstance._rootNodeID ?
+      this._reactInternalInstance._rootNodeID :
+      this._reactInternalInstance._nativeContainerInfo._idCounter;
+    this.uidGenerator = this.uidGenerator || new UIDGenerator(seed);
+    return this.uidGenerator;
+  }
+
   render() {
 
-    const type = this.props.type;
-    const options = this.props.options || nooptions;
     const stylesheet = this.props.stylesheet || Form.stylesheet;
     const templates = this.props.templates || Form.templates;
     const i18n = this.props.i18n || Form.i18n;
 
-    t.assert(t.isType(type), `[${SOURCE}] missing required prop type`);
-    t.assert(t.Object.is(options), `[${SOURCE}] prop options must be an object`);
-    t.assert(t.Object.is(stylesheet), `[${SOURCE}] missing stylesheet config`);
-    t.assert(t.Object.is(templates), `[${SOURCE}] missing templates config`);
-    t.assert(t.Object.is(i18n), `[${SOURCE}] missing i18n config`);
+    if (process.env.NODE_ENV !== 'production') {
+      t.assert(t.isType(this.props.type), `[${SOURCE}] missing required prop type`);
+      t.assert(t.maybe(t.Object).is(this.props.options) || t.Function.is(this.props.options) || t.list(t.maybe(t.Object)).is(this.props.options), `[${SOURCE}] prop options, if specified, must be an object, a function returning the options or a list of options for unions`);
+      t.assert(t.Object.is(stylesheet), `[${SOURCE}] missing stylesheet config`);
+      t.assert(t.Object.is(templates), `[${SOURCE}] missing templates config`);
+      t.assert(t.Object.is(i18n), `[${SOURCE}] missing i18n config`);
+    }
 
-    const Component = getComponent(type, options);
+    const value = this.props.value;
+    const type = getTypeFromUnion(this.props.type, value);
+    const options = getComponentOptions(this.props.options, noobj, value, this.props.type);
 
-    return React.createElement(Component, {
+    // this is in the render method because I need this._reactInternalInstance._rootNodeID in React ^0.14.0
+    // and this._reactInternalInstance._nativeContainerInfo._idCounter in React ^15.0.0
+    const uidGenerator = this.getUIDGenerator();
+
+    return React.createElement(getFormComponent(type, options), {
       ref: 'input',
       type: type,
       options: options,
@@ -550,6 +786,7 @@ class Form extends React.Component {
       onChange: this.props.onChange || noop,
       ctx: {
         context: this.props.context,
+        uidGenerator,
         auto: 'labels',
         stylesheet,
         templates,
@@ -562,12 +799,13 @@ class Form extends React.Component {
 }
 
 module.exports = {
-  getComponent,
+  getComponent: getFormComponent,
   Component,
   Textbox,
   Checkbox,
   Select,
   DatePicker,
   Struct,
+  List: List,
   Form
 };
