@@ -1,4 +1,8 @@
 module.exports = {
   optional: ' (optional)',
-  required: ''
+  required: '',
+  add: 'Add',
+  remove: '✘',
+  up: '↑',
+  down: '↓'
 };
