@@ -3,5 +3,6 @@ module.exports = {
   checkbox: require('./checkbox'),
   select: require('./select'),
   datepicker: require('./datepicker'),
-  struct: require('./struct')
+  struct: require('./struct'),
+  list: require('./list')
 };
