@@ -1,10 +1,5 @@
 'use strict';
 
-var mockery = require('mockery');
-mockery.enable();
-mockery.warnOnUnregistered(false);
-mockery.registerSubstitute('react-native', 'react');
-
 var tape = require('tape');
 var t = require('tcomb-validation');
 var bootstrap = {
@@ -12,28 +7,36 @@ var bootstrap = {
   datepicker: function () {},
   select: function () {},
   struct: function () {},
-  textbox: function () {}
+  textbox: function () {},
+  list: function () {}
 };
 
 var core = require('../lib/components');
+import { UIDGenerator } from '../lib/util';
 
 var ctx = {
   auto: 'labels',
   label: 'ctxDefaultLabel',
   templates: bootstrap,
-  i18n: {optional: ' (optional)', required: ''}
+  i18n: {optional: ' (optional)', required: ''},
+  uidGenerator: new UIDGenerator('seed'),
+  path: []
 };
 var ctxPlaceholders = {
   auto: 'placeholders',
   label: 'ctxDefaultLabel',
   templates: bootstrap,
-  i18n: {optional: ' (optional)', required: ''}
+  i18n: {optional: ' (optional)', required: ''},
+  uidGenerator: new UIDGenerator('seed'),
+  path: []
 };
 var ctxNone = {
   auto: 'none',
   label: 'ctxDefaultLabel',
   templates: bootstrap,
-  i18n: {optional: ' (optional)', required: ''}
+  i18n: {optional: ' (optional)', required: ''},
+  uidGenerator: new UIDGenerator('seed'),
+  path: []
 };
 
 tape('Textbox', function (tape) {
@@ -1120,5 +1123,52 @@ tape('DatePicker', function (tape) {
 
 });
 
-mockery.deregisterAll();
-mockery.disable();
+tape('List', ({ test }) => {
+
+  var List = core.List;
+
+  test('should support unions', (assert) => {
+    assert.plan(2)
+
+    const AccountType = t.enums.of([
+      'type 1',
+      'type 2',
+      'other'
+    ], 'AccountType')
+
+    const KnownAccount = t.struct({
+      type: AccountType
+    }, 'KnownAccount')
+
+    const UnknownAccount = KnownAccount.extend({
+      label: t.String,
+    }, 'UnknownAccount')
+
+    const Account = t.union([KnownAccount, UnknownAccount], 'Account')
+
+    Account.dispatch = value => value && value.type === 'other' ? UnknownAccount : KnownAccount
+
+    let component = new List({
+      type: t.list(Account),
+      ctx: ctx,
+      options: {},
+      value: [
+        { type: 'type 1' }
+      ]
+    })
+
+    assert.strictEqual(component.getItems()[0].input.props.type, KnownAccount)
+
+    component = new List({
+      type: t.list(Account),
+      ctx: ctx,
+      options: {},
+      value: [
+        { type: 'other' }
+      ]
+    })
+
+    assert.strictEqual(component.getItems()[0].input.props.type, UnknownAccount)
+  })
+})
+
